Added Dockerfile based build
Added a Dockerfile that does cenntos build to make it easier for new people to try things out. The image is a bit too heavy weight (~850 MB). Tried using alpine, but ran into fortran compile errors that seemed annoying to have to fix when it already works on couple of other distributions.